UtilsText, Registry: pass buffer sizes as size_t data type

+++ Common/Header/UtilsText.hpp
@@ -46,8 +46,8 @@ Copyright_License {
  */
 void StringChomp(TCHAR *p);
 
-bool ReadString(ZZIP_FILE* zFile, int Max, TCHAR *String);
-bool ReadStringX(FILE *fp, int Max, TCHAR *String);
+bool ReadString(ZZIP_FILE* zFile, size_t Max, TCHAR *String);
+bool ReadStringX(FILE *fp, size_t Max, TCHAR *String);
 
 double StrToDouble(const TCHAR *Source, const TCHAR **Stop);
 void PExtractParameter(const TCHAR *Source, TCHAR *Destination,
